SAS-332 Hecha la división de los tests de manejo de efectos en base a lo colocado en la descripción de la tarjeta del Jira
