+++ tests/game_logic_tests/effects_tests/__init__.py
@@ -1,0 +1,1 @@
+from .. import *  # noqa : F401
